+++ CHANGELOG.md
@@ -1,5 +1,9 @@
+# Changelog
+
+> This file was generated automatically using [@semantic-release](https://github.com/semantic-release/semantic-release).
+
 ## 1.0.0 (2024-09-14)
 
 ### Features
 
-* initial release ([6969503](https://github.com/zavoloklom/docker-compose-linter/commit/69695032957556141669ea6a5daf213ba8479ffa))
+- initial release ([6969503](https://github.com/zavoloklom/docker-compose-linter/commit/69695032957556141669ea6a5daf213ba8479ffa))